Reorganize contracts module
