+++ populus/contracts/__init__.py
@@ -1,0 +1,8 @@
+# NOQA
+from populus.contracts.utils import (  # NOQA
+    get_max_gas,
+    deploy_contract,
+)
+from populus.contracts.functions import Function  # NOQA
+from populus.contracts.events import Event  # NOQA
+from populus.contracts.core import Contract  # NOQA
